+++ NEWS.md
@@ -13,6 +13,7 @@ Other:
 * `print.marginaleffects` now prints all columns supplied to `newdata`
 * Informative error when `hypothesis` includes a `b#` larger than the available number of estimates.
 * `avg_predictions(model, variables = "x")` computes average counterfactual predictions by subgroups of `x`
+* `hypotheses()` now works on lists and in calls to `lapply()`, `purrr::map()`, etc.
 
 Bugfixes: